Improvements to search, some tv support added
Fonts updated for detail layouts
Memory leak improvements
Removed Android ViewModel using a simple custom view model as a base class
Fixed a bug on the actor screen when a cached actor would show before the expected actor

+++ Movies/moviedb/src/main/java/com/dev/cameronc/moviedb/data/movie/detail/MovieDetailsResponse.kt
@@ -16,7 +16,7 @@ data class MovieDetailsResponse(
         @SerializedName("homepage")
     val homepage: String,
         @SerializedName("id")
-    val id: Int,
+    val id: Long,
         @SerializedName("imdb_id")
     val imdbId: String,
         @SerializedName("original_language")
@@ -38,7 +38,7 @@ data class MovieDetailsResponse(
         @SerializedName("release_dates")
     val releaseDates: ReleaseDates,
         @SerializedName("revenue")
-    val revenue: Int,
+    val revenue: Long,
         @SerializedName("runtime")
     val runtime: Int,
         @SerializedName("spoken_languages")
